slurm: improve script; exit if we fail to initialise the GPU in tensorflow
The latter is optional and implemented as a flag --only-gpu that we 
specify in the slurm job script. The reason for this if we for whatever 
reason fail to initialise the GPU, it'll use the CPU instead 
automatically - which would result in us running out of RAM very quickly 
and potentially affecting other users of the system.

With these changes we now *shouldn't* need the --exclusive flag anymore.